#4795 disable menu item if database not connected (#4828)
* #4795 disable menu item if database not connected

* resolve merge conflict

* resolve merge conflict

* resolve merge conflicts

* resolve check style

* fix import order

* resolve merge

* delete duplicate code

* change push to application implementation

* fix check style

* delete action

* put deprecated annotation on OldDatabaseCommandWrapper
