+++ src/main/java/org/jabref/gui/push/PushToApplicationAction.java
@@ -1,39 +1,79 @@
 package org.jabref.gui.push;
 
-import java.awt.event.ActionEvent;
 import java.util.List;
 import java.util.Optional;
 
-import javax.swing.AbstractAction;
-import javax.swing.Action;
 import javax.swing.SwingUtilities;
 
+import org.jabref.Globals;
 import org.jabref.JabRefExecutorService;
 import org.jabref.gui.BasePanel;
 import org.jabref.gui.JabRefFrame;
+import org.jabref.gui.StateManager;
+import org.jabref.gui.actions.Action;
+import org.jabref.gui.actions.SimpleCommand;
+import org.jabref.gui.icon.JabRefIcon;
+import org.jabref.gui.keyboard.KeyBinding;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.model.entry.BibEntry;
+import org.jabref.preferences.JabRefPreferences;
+
+import org.fxmisc.easybind.EasyBind;
 
 /**
  * An Action class representing the process of invoking a PushToApplication operation.
  */
-class PushToApplicationAction extends AbstractAction implements Runnable {
+public class PushToApplicationAction extends SimpleCommand implements Runnable {
 
-    private final PushToApplication operation;
-    private final JabRefFrame frame;
+    private PushToApplication operation;
+    private JabRefFrame frame;
     private BasePanel panel;
     private List<BibEntry> entries;
 
-    public PushToApplicationAction(JabRefFrame frame, PushToApplication operation) {
+    public PushToApplicationAction(final JabRefFrame frame, final StateManager stateManager) {
         this.frame = frame;
-        putValue(Action.SMALL_ICON, operation.getIcon());
-        putValue(Action.NAME, operation.getName());
-        putValue(Action.SHORT_DESCRIPTION, operation.getTooltip());
-        this.operation = operation;
+        this.operation = getLastUsedApplication(frame.getPushApplications().getApplications());
+        this.executable.bind(EasyBind.map(stateManager.activeDatabaseProperty(), Optional::isPresent));
+    }
+
+    private PushToApplication getLastUsedApplication(List<PushToApplication> pushActions) {
+        String appSelected = Globals.prefs.get(JabRefPreferences.PUSH_TO_APPLICATION);
+        for (PushToApplication application : pushActions) {
+            if (application.getApplicationName().equals(appSelected)) {
+                return application;
+            }
+        }
+
+        // Nothing found, pick first
+        return pushActions.get(0);
+    }
+
+    public Action getActionInformation() {
+        return new Action() {
+            @Override
+            public Optional<JabRefIcon> getIcon() {
+                return Optional.of(operation.getIcon());
+            }
+
+            @Override
+            public Optional<KeyBinding> getKeyBinding() {
+                return Optional.of(KeyBinding.PUSH_TO_APPLICATION);
+            }
+
+            @Override
+            public String getText() {
+                return Localization.lang("Push entries to external application (%0)", operation.getApplicationName());
+            }
+
+            @Override
+            public String getDescription() {
+                return "";
+            }
+        };
     }
 
     @Override
-    public void actionPerformed(ActionEvent e) {
+    public void execute() {
         panel = frame.getCurrentBasePanel();
 
         // Check if a BasePanel exists:
@@ -44,7 +84,7 @@ class PushToApplicationAction extends AbstractAction implements Runnable {
         // Check if any entries are selected:
         entries = panel.getSelectedEntries();
         if (entries.isEmpty()) {
-            frame.getDialogService().showErrorDialogAndWait((String) getValue(Action.NAME),
+            frame.getDialogService().showErrorDialogAndWait(operation.getApplicationName(),
                     Localization.lang("This operation requires one or more entries to be selected."));
 
             return;
@@ -54,7 +94,7 @@ class PushToApplicationAction extends AbstractAction implements Runnable {
         if (operation.requiresBibtexKeys()) {
             for (BibEntry entry : entries) {
                 if (!(entry.getCiteKeyOptional().isPresent()) || entry.getCiteKeyOptional().get().trim().isEmpty()) {
-                    frame.getDialogService().showErrorDialogAndWait((String) getValue(Action.NAME),
+                    frame.getDialogService().showErrorDialogAndWait(operation.getApplicationName(),
                             Localization.lang("This operation requires all selected entries to have BibTeX keys defined."));
 
                     return;
